Jan 23, 2025, 10:43 AM

+++ CT07_03/lesson3.py
@@ -50,6 +50,6 @@
 
 savings = 0
 while savings < 100:
-    savings += int(input("How much did you save today? "))
+    savings += float(input("How much did you save today? "))
     print("You now have $" + str(savings) + ".")
 print("You saved $100 or more. ($" + str(savings) + ") Congrats!")
